Clean Install autoattended
Remove some apt's updates. Merge dependencies in one apt install. A question for a LAN or WAN server for bypass the install of certbot. Add a root checker and exits if it's in a normal environment. Add support for Debian 9 and 10. Detects if the files are downloaded and bypass the download.