fix: Prevent invalid authorization headers from accessing the API
The API is secured by this authorization filter that should prevent access to the API when the auhorization header is invalid. Instead the filter just set the status to 403, but delivered the data anyway.

This commit fixes this by explicitly setting the result to an UnauthorizedResult, which prevents access to the API.

#495